Feat/initial SummaryPage setup (#23)
* feat: Created SummaryPage

Created a basic Summarypage, with dummy info. Contains information about the interview and duration.

* feat: Linked Summary with Interview page

Linked SummaryPage.js with the InterviewPractice.js page, by adding a Next button. Also changed implementation of time storage in InterviewPractice.js, to allow it to pass on stored state and display it on SummaryPage.

* feat: Change home button color

Changed the color of the home button to prevent it from seeming disabled (to users)